Separated main app file into file for each page
